TD2 - Partie 3
Added : 
file opening and saving, using two formats ".ch" for cyphered texts and ".dech" for decyphered texts. 

Fixed :
the method .at() for text was replaced, it would cause the program to crash sometimes, and it was not clear why.

Special char, like é,è,ê... are still a problem and cannot be decyphered yet.